Make attained_service a 64 bit unsigned int to be extremely safe.

+++ flow_queue.h
@@ -1,7 +1,7 @@
 struct onramp_flow_queue {
 	struct sk_buff	  *head;
 	struct sk_buff	  *tail;
-	int		  attained_service;
+	u64		  attained_service;
 }; /* please try to keep this structure <= 64 bytes */
 
 /* Helper functions : might be changed when/if skb use a standard list_head */
